Move addition of versioning to submodule
